### 2019-02-16 #1246 @labarba
👋  @gkthiruvathukal — This JOSS submission looks like something you could handle as editor, no? I will assign you, but let me know if I can help you in any way!
