### 2019-03-04 #1246 @arfon
:wave: @gkthiruvathukal - how are you getting along identifying reviewers? @JedS6391 - if you get a chance, identifying a few potential reviewer names from the list (https://bit.ly/joss-reviewers) would help out too.
