### 2019-03-13 #1246 @gkthiruvathukal
@arfon I'm getting the ball rolling now. Thanks for everyone's patience as I begin my tenure with JOSS.

@JedS6391 Can you let me know if you have any preferred reviewers? We do have reviewers with Kotlin expertise. As I wish to get the review processing moving as soon as possible, if I do not hear any preferences, I will make the assignment within 24-48 hours of this posting.
