### 2019-03-14 #1246 @gkthiruvathukal
@paragkulkarni11 Would you be willing to contribute a review for JOSS on this submission? This submission seems closely related to your areas of expertise and interest.
