### 2019-03-15 #1246 @gkthiruvathukal
@arcuri82 Would you be willing to contribute a review for JOSS on this submission? This submission seems closely related to your areas of expertise and interest. I have not seen a follow-up communication from @paragkulkarni11 yet. If either or both of you respond, we can begin the review process.
