### 2019-03-19 #1246 @gkthiruvathukal
@homonecloco Would you be willing to contribute a review to this JOSS submission? I see you have Java and Genetics listed. I realize that genetic programming is not the same thing as genetics, but your expertise would be helpful for this submission.
