### 2019-03-20 #1337 @whedon
[REVIEW]: LGP: A robust Linear Genetic Programming implementation on the JVM using Kotlin.
**Submitting author:** @JedS6391 (<a href="http://orcid.org/0000-0002-5028-9987">Jed Simson</a>)
**Repository:** <a href="https://github.com/JedS6391/LGP" target ="_blank">https://github.com/JedS6391/LGP</a>
**Version:** 5.1

- [x] **Version:** 5.1
