### 2019-04-01 JedS6391/LGP#55 @homonecloco
Broken links in documentation
The following links are broken:

 *  https://lgp.readthedocs.io/en/latest/guide/environment.html from https://lgp.readthedocs.io/en/latest/guide/environment.html
 * https://lgp.jedsimson.co.nz/api/html/lgp.core.evolution.population/-evolution-model/index.html from https://lgp.readthedocs.io/en/latest/guide/evolution.html
* https://jeds6391.github.io/LGP/api/html/lgp.core.evolution/index.html from https://lgp.readthedocs.io/en/latest/guide/problem.html

Please review all the links to the API. They seem to be consistently broken
