### 2019-04-01 #1337 @homonecloco
The documentation is extensive and detailed. However, the links to the API are broken. Reported as an issue in JedS6391/LGP#55
