### 2019-04-01 #1337 @homonecloco
There is a link to the Thesis of the author, http://www.jedsimson.co.nz/static/assets/thesis/Open-Source%20Linear%20Genetic%20Programming.pdf Would it be appropriate to also include the supervisor as author?
