### 2019-04-01 JedS6391/LGP#56 @homonecloco
Installation issues
The required version of the JDK is not stated in the documentation.  This can be resolved by just stating the java version in the documentation (I had to update to java 1.8, as I had 1.7 installed).

I'm following the instructions in https://github.com/JedS6391/LGP/tree/release/4.2

However, kaitlin is not compiling the ```MyProble.kt```. I get the following issue:

```sh
$ java -version
java version "1.8.0_201"
Java(TM) SE Runtime Environment (build 1.8.0_201-b09)
Java HotSpot(TM) 64-Bit Server VM (build 25.201-b09, mixed mode)
$ kotlinc -cp LGP.jar -no-jdk -no-stdlib MyProblem.kt
MyProblem.kt:1:36: error: unresolved reference: Configuration
import lgp.core.environment.config.Configuration
                                   ^
MyProblem.kt:11:22: error: unresolved reference: Config
            config = Config()
                     ^
$ ls
10.21105.joss.01337.pdf	LGP.jar			MyProblem.kt
```
