### 2019-04-01 #1337 @homonecloco
I had issues running the program, as stated JedS6391/LGP#56 . I'll continue the review once this issue is resolved.
