### 2019-04-04 2df3248
Fix more broken links in docs (#55) and add JDK version to README (#56).
