+++ docs/source/guide/programs.rst
@@ -10,9 +10,13 @@ Despite the fairly unrestricted interface of the program modules, LGP describes 
 
 The ``Program`` interface is designed this way in the name of flexibility; instead of restricting the shape and operation of the component, the individual program logic can be customised to allow for situations where simply executing the instructions is not enough. For example, a control problem may involve moving a robot through an environment. In this case, instructions might be commands such as move left or rotate 90°, and the program would be responsible for gathering input data from the robot's environment and providing it to the instructions.
 
-The built-in program implementation (``lgp.lib.BaseProgram``), provides a simple, single-output program interface. Instructions of the program are executed in sequential order and there is support for branching (i.e. conditional instructions). Furthermore, the ``findEffectiveProgram()`` method implements the intron elimination algorithm as outlined by Brameier, M. F., & Banzhaf, W. (2007) [#f1]_.
+The built-in program implementation (``BaseProgram``), provides a simple program interface which can handle both single-output and multiple-output scenarios, depending on the output resolver used.
+
+Instructions of the program are executed in sequential order and there is support for branching (i.e. conditional instructions). Furthermore, the ``findEffectiveProgram()`` method implements the intron elimination algorithm as outlined by Brameier, M. F., & Banzhaf, W. (2007) [#f1]_.
 
 The interaction between instructions and programs can be leveraged through the modular interface to adapt the system to the needs of an individual problem.
+
+.. note:: The ``BaseProgram`` type is found in the ``nz.co.jedsimson.lgp.lib`` package which resides in the `LGP-lib repository <https://github.com/JedS6391/LGP-lib>`_.
 
 Program Generator
 =================
@@ -21,7 +25,14 @@ Program generation is facilitated by the ``ProgramGenerator`` class. Again, in t
 
 The program generator is expected to act as a stream of programs so that other components in the system can continuously generate new programs.
 
-The built-in program generator (``lgp.lib.BaseProgramGenerator``), creates a random, endless stream of effective programs. Properties of these initial programs can be tuned through the environment.
+Two built-in program generators are offered:
+
+1. ``RandomProgramGenerator`` creates a random, endless stream of programs.
+2. ``EffectiveProgramGenerator`` creates a random, endless steam of effective programs (i.e. they are guaranteed to have an effect on the output register(s)).
+
+Properties of the initial programs generated by each implementation can be tuned through the environment definition.
+
+.. note:: The ``BaseProgram`` type is found in the ``nz.co.jedsimson.lgp.lib`` package which resides in the `LGP-lib repository <https://github.com/JedS6391/LGP-lib>`_.
 
 Program Translator
 ==================
@@ -48,8 +59,8 @@ The output from ``BaseProgramTranslator`` should compile under most systems usin
 API
 ===
 
-See `lgp.core.evolution.population <https://jeds6391.github.io/LGP/api/html/lgp.core.evolution.population/index.html>`_ for information on the ``Program`` and ``ProgramGenerator`` interfaces.
+See `nz.co.jedsimson.lgp.core.program <https://lgp.jedsimson.co.nz/api/html/nz.co.jedsimson.lgp.core.program/index.html>`_ for information on the ``Program``, ``ProgramGenerator`` and ``ProgramTranslator`` APIs.
 
-API details for the built-in implementations are found in `lgp.lib <https://jeds6391.github.io/LGP/api/html/lgp.lib/index.html>`_.
+API details for the built-in implementations are found in `nz.co.jedsimson.lgp.lib <https://github.com/JedS6391/LGP-lib>`_.
 
 .. [#f1] Brameier, M. F., & Banzhaf, W. (2007). Linear Genetic Programming. Springer Science & Business Media. https://doi.org/10.1007/978-0-387-31030-5
