### 2019-04-04 JedS6391/LGP#56 @JedS6391
Hi @homonecloco,

Firstly, thanks for reviewing LGP for JOSS!

The documentation has been updated on the [master branch](https://github.com/JedS6391/LGP/tree/master) to reflect these API changes and should work as intended now.

Alternatively, the instructions in the [usage guide](https://lgp.readthedocs.io/en/latest/guide/usage.html) have some slightly more detailed steps for setting up the same example problem.

I have also added a note to both the README and the usage guide to indicate the required JDK/java version.

Please let me know if you are able to get the sample working correctly after these changes.
