### 2019-04-04 JedS6391/LGP#55 @JedS6391
Hi @homonecloco, I have updated the links across the [documentation](https://lgp.readthedocs.io/en/latest/). I believe all the links should be OK now, but please let me know if I have missed any.

I also reviewed the documentation content and made some minor adjustments where I felt it was necessary.
