### 2019-04-06 JedS6391/LGP#56 @homonecloco
Thanks for the updates. I managed to make it run, but I think in the manual you still have something inconsisten, as I was unable to run the program with:

```sh
kotlinc -cp LGP.jar -no-jdk -no-stdlib MyProblem.kt
kotlin -cp LGP.jar:. MyProblemKt
```
I think it is because even renaming the LGP-core to LGP.jar is not enough.

However, I did succeed to run

```sh
# Compile to a JAR using kotlinc
kotlinc -cp LGP-core.jar:LGP-lib.jar -no-jdk -no-stdlib -d MyProblem.jar MyProblem.kt

# Use the Kotlin interpreter to execute the main function
kotlin -cp LGP-core.jar:LGP-lib.jar:MyProblem.jar:. MyProblemKt
```
