### 2019-04-06 b154c43
Make compilation and execution steps consistent between docs.
