+++ README.md
@@ -102,7 +102,7 @@ kotlinc -cp LGP-core.jar:LGP-lib.jar -no-jdk -no-stdlib MyProblem.kt
 This will generate a class file in the directory called `MyProblemKt.class`. To interpret the class file using the Kotlin interpreter is simple:
 
 ```bash
-kotlin -cp LGP.jar:LGP-lib.jar:. MyProblemKt
+kotlin -cp LGP-core.jar:LGP-lib.jar:. MyProblemKt
 ```
 
 You should see the following output:
