+++ docs/source/guide/usage.rst
@@ -68,13 +68,13 @@ To compile, we use ``kotlinc``:
 
 .. code-block:: bash
 
-    kotlinc -cp LGP.jar -no-jdk -no-stdlib MyProblem.kt
+    kotlinc -cp LGP-core.jar:LGP-lib.jar -no-jdk -no-stdlib MyProblem.kt
 
 This will generate a class file in the directory called ``MyProblemKt.class``. To interpret the class file using the Kotlin interpreter is simple:
 
 .. code-block:: bash
 
-    kotlin -cp LGP.jar:. MyProblemKt
+    kotlin -cp LGP-core.jar:LGP-lib.jar:. MyProblemKt
 
 You should see the following output:
 
