### 2019-04-06 JedS6391/LGP#56 @JedS6391
Apologies, @homonecloco!

Looks like I had missed a few updates. These inconsistencies should be resolved now with the latest commit on the `master` branch (b154c4313ed6f8d2a7eae06ee72cc922bfb774eb).
