### 2019-04-07 #1337 @homonecloco
The installation consists on downloading the jars. There is also an option to do it from Maven.
