### 2019-04-07 JedS6391/LGP#57 @homonecloco
Unit testings are missing. 
As per the checklist in openjournals/joss-reviews#1337 , a set of Unit tests is required. I see that travis is used to verify that the system is building, but you could add also your unit tests. You can use JUnit to make sure that all the basics operations are covered.
