### 2019-04-22 #1337 @JedS6391
Hi @gkthiruvathukal, [JedS6391/LGP#55](https://github.com/JedS6391/LGP/issues/55) and [JedS6391/LGP#56](https://github.com/JedS6391/LGP/issues/56) have been addressed. I am currently working on [JedS6391/LGP#7](https://github.com/JedS6391/LGP/issues/57), but it is taking a little while to write the unit tests before the issue can be closed.
