### 2019-05-02 #1337 @gkthiruvathukal
Thanks, @JedS6391. Please let us know when you've had a chance to add unit tests.
