### 2019-06-07 #1337 @JedS6391
Much appreciated! I will able to start working on the remaining issues over the next few weekends.
