### 2019-06-17 8783100
Run tests during build.
