+++ .travis/build.sh
@@ -2,7 +2,13 @@
 
 chmod u+x gradlew
 
+echo "Compiling core JAR..."
+
 ./gradlew coreJar --no-daemon --stacktrace --console=plain
+
+echo "Running tests..."
+
+./gradlew test --no-daemon --info --console=plain
 
 echo "real branch:"
 echo $REAL_BRANCH
