### 2019-06-18 #1337 @JedS6391
Hi @gkthiruvathukal -- just wanted to give you a brief progress update.

I have implemented unit tests for a subset of the functionality in [this PR](https://github.com/JedS6391/LGP/pull/58). I hope to finish the remaining tests over the weekend.

I will let you know once all unit tests have been completed.
