### 2019-06-19 a1b998e
Generate test report and code coverage during build.
