### 2019-06-19 7138005
Add diagnostics to build script.
