+++ .travis/build.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env sh
+#!/bin/bash
 
 chmod u+x gradlew
 
@@ -14,7 +14,11 @@ echo "Generating coverage report..."
 
 ./gradlew jacocoTestReport
 
+echo "Uploading report to codecov.io..."
+
 bash <(curl -s https://codecov.io/bash) -t $CODECOV_TOKEN
+
+echo "Complete"
 
 echo "real branch:"
 echo $REAL_BRANCH
