### 2019-06-19 cf39bd9
Add code coverage badge to README.
