### 2019-06-19 #1337 @JedS6391
Hi @gkthiruvathukal, the latest [commit](https://github.com/JedS6391/LGP/commits/develop) against the develop branch has a green build and has a subset of the tests.

At the moment, the coverage is 20% but I intended to increase this going forward. If you are interested to see a breakdown of the areas covered, then that can be viewed [here](https://codecov.io/gh/JedS6391/LGP/tree/8249ba28013930c7ca9b1328a715e64d83960d4f/src/main/kotlin/nz/co/jedsimson/lgp/core).

I will continue to add more tests and eventually close the initial issue.
