### 2019-06-24 4544839
Update code coverage badge URL.
