### 2019-06-26 24acb85
Update changelog and version number.
