+++ src/main/kotlin/nz/co/jedsimson/lgp/core/modules/README.md
@@ -1,0 +1,19 @@
+# Package nz.co.jedsimson.lgp.core.modules
+
+## About
+
+The `nz.co.jedsimson.lgp.core.modules` API provides two main components:
+  
+  - Interfaces to define a module in the system, as well as metadata about modules:
+    - `Module`
+    - `ModuleInformation`
+  - Definitions for managing collections of `Module`s and creating instance of a `Module`.
+    - `ModuleContainer`
+    - `ModuleFactory`
+    
+`Module`s are core to the system, as any number of them can be registered and then used in different parts 
+of the system. This extensible design allows control of all the aspects of the system. 
+
+The goal of the `nz.co.jedsimson.lgp.core` package is to provide a base set of modules for general purpose
+usage of the system for evolving programs tailored to solve particular problems, and offering extension points
+through the modules API.
