### 2019-07-14 3786589
Add events infrastructure and tests. Add initial per-module documentation.
