### 2019-07-18 #1337 @gkthiruvathukal
@JedS6391 I find your efforts satisfactory as there is now a basis for doing further testing in future work. If our reviewer, @homonecloco, agrees, I would like to move toward accepting your submission. Once we hear from @homonecloco, you will need to create a Zenodo archive.
