### 2019-07-20 3543655
Add further module documentation.
