### 2019-07-23 #1337 @gkthiruvathukal
@homonecloco When you get a chance, can you let us know if your issues are all addressed so I can proceed with the next steps (acceptance).
