### 2019-07-23 #1337 @gkthiruvathukal
@homonecloco No problem! Thanks for your quick response. End of week is fine.
