### 2019-08-01 #1337 @gkthiruvathukal
@homonecloco Any chance you can let us know whether your remaining issues have been addressed so I can proceed with acceptance?
