### 2019-08-24 25b4469
Use OpenJDK 8 instead of Oracle JDK.
