+++ .travis.yml
@@ -3,17 +3,13 @@ matrix:
   include:
   - os: linux
     sudo: false
-    jdk: oraclejdk9
-    addons:
-      apt:
-        packages:
-        - oracle-java9-installer
-        - oracle-java9-set-default
+
 env:
   global:
     # Because release builds work with tagged commits (which in our case will always be related to a branch),
     # we need to attempt to fetch the real branch name for a tagged commit.
     - REAL_BRANCH=$(git ls-remote origin | sed -n "\|$TRAVIS_COMMIT\s\+refs/heads/|{s///p}")
+    - JAVA_HOME=/usr/lib/jvm/java-8-openjdk-amd64
 before_cache:
   - rm -f  $HOME/.gradle/caches/modules-2/modules-2.lock
   - rm -fr $HOME/.gradle/caches/*/plugin-resolution/
@@ -22,6 +18,10 @@ cache:
   - $HOME/.gradle/caches/
   - $HOME/.gradle/wrapper/
 before_install:
+  - sudo add-apt-repository -y ppa:openjdk-r/ppa
+  - sudo apt-get -qq update
+  - sudo apt-get install -y openjdk-8-jdk --no-install-recommends
+  - sudo update-java-alternatives -s java-1.8.0-openjdk-amd64
   - chmod -R ug+x .travis
 script:
   - chmod -R ug+x .travis
