### 2019-08-24 d68e8b4
Suppress Gradle welcome message.
