### 2019-09-01 #1337 @homonecloco
Hi,
Sorry for not finishing the review earlier. I ended up having a couple of complex weeks. The unit tests where added and they run. They seem to be ok. There is also an extra repository with actual examples, that may be good to have included in the main page of the project. There is one example that is missing to run, so I aded an issue on the corresponding repository:
JedS6391/LGP-examples#1 and I would like a bit more clarification as stated in JedS6391/LGP-examples#2
But other than that, I'm happy with the state of the project.
Best,
Ricardo.
