### 2019-10-15 #1337 @gkthiruvathukal
Thanks for the nudge, @Kevin-Mattheus-Moerman.

@openjournals/joss-eics  I am ready to recommend acceptance of the submission. @JedS6391 has done a great job of addressing all reviewer feedback during this process.
