### 2019-10-16 #1337 @Kevin-Mattheus-Moerman
Here are some additional points in relation to your paper:

- [ ] We will shortly process the acceptance of this work in JOSS. Please double check your work (including the author list, affiliations, and acknowledgements) and make any final edits now.
- [ ] Please check your reference syntax throughout and make sure it is in line with these guidelines:
https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax
e.g. you use `[@friedberg1958] experimented with...` in your paper.md, but this should be `@friedberg1958 experimented with...` if used in this way.
- [ ] The final paper will contain a link to the archived software. I therefore suggest removal of this sentence and the links to the Zenodo archives "The source code for LGP and LGP-lib have been archived to Zenodo with the linked DOIs: • LGP core: (Simson, 2019a) • LGP library: (Simson, 2019b)". If both the core and the library are important (and were reviewed), does it make sense to place them both in a single archive?
- [ ] Once you've made the above changes please archive a copy of the reviewed software on Zenodo and report the DOI here.
- [ ] Can you confirm what the latest version tag is of the reviewed software? Is it still 4.2 or was the version tag altered during review?

Thanks.
