### 2019-10-17 #1337 @JedS6391
Hi @Kevin-Mattheus-Moerman, thanks for the notes. I will make the request changes and update this thread once all changes are completed.

Regarding your last question -- the version tag was altered during review to resolve some issues that were encountered. The latest version tag (and reviewed version) is [5.1](https://github.com/JedS6391/LGP/tree/5.1). This will be the version that I archive on Zenodo.

Thanks!
