### 2019-10-22 7d3890e
Update JOSS paper based on feedback.
