### 2019-10-22 #1337 @JedS6391
Hi @Kevin-Mattheus-Moerman, I have made the requested changes to the paper. The DOI is 10.5281/zenodo.3402305 (the archive can be found [here](https://zenodo.org/record/3402305)).
