### 2019-10-22 #1337 @Kevin-Mattheus-Moerman
@JedS6391 can you modify the archived version on Zenodo so that the title (and author list) match the paper?
