### 2019-10-22 #1337 @Kevin-Mattheus-Moerman
@JedS6391 I cannot see the changes yet (see boxes left  unticked, the references style is still the same and the section referring to the source is still there/unaltered)? Has the paper.md file been updated? Is the above pdf accurate?
