### 2019-10-25 #1337 @JedS6391
Hi @Kevin-Mattheus-Moerman, I think something went wrong with the automatic upload to Zenodo from GitHub so I created the archive manually (see [here](https://zenodo.org/record/3519400)). I have also updated the archive title + author list to be correct.

The new DOI is 10.5281/zenodo.3519400.

Apologies for the delay in getting this sorted.
