### 2019-10-25 #1337 @JedS6391
I believe I need to change the DOI to the new one to get the paper to reflect the changes made?
