### 2019-10-26 #1337 @Kevin-Mattheus-Moerman
@JedS6391 no you need to make the changes in your paper.md file (https://github.com/JedS6391/LGP/blob/develop/paper.md) and run '@whedon generate pdf'. Note the two changes I proposed above. Let me know if you have questions.
