### 2019-10-26 #1337 @JedS6391
Hi @Kevin-Mattheus-Moerman, the generated PDF now has the reflected changes (I didn't realise whedon read from the `develop` branch). Please note that the correct DOI is as stated yesterday: 10.5281/zenodo.3519400.
